server-bak/requirements.txt
Flask
earthengine-api
google-cloud-firestore
google-cloud-pubsub
oauth2client
mercantile
tensorflow
